refactor(#1): 디렉토리 이름 변경 - error보다는 exception이 더 적합하여 이름을 변경했어요.
